+++ config/routes.rb
@@ -34,6 +34,7 @@ Rails.application.routes.draw do
     get :downloads_by_month, path: "downloads-by-month"
     get :downloads_by_quarter, path: "downloads-by-quarter"
     get :agreement_reports, path: "agreement-reports"
+    get :searches
     resources :replies, only: :index
     resources :tools
     root action: :dashboard